+++ crates/lib/src/games/minecraft/protocol/legacy_vb1_8.rs
@@ -1,11 +1,11 @@
 use crate::{
     buffer::{Buffer, Utf16Decoder},
     games::minecraft::{JavaResponse, LegacyGroup, Server},
+    protocols::types::TimeoutSettings,
     socket::{Socket, TcpSocket},
     utils::{error_by_expected_size, retry_on_timeout},
     GDErrorKind::{PacketBad, ProtocolFormat},
     GDResult,
-    TimeoutSettings,
 };
 
 use std::net::SocketAddr;
