+++ crates/lib/src/games/theship.rs
@@ -1,12 +1,11 @@
 use crate::{
     protocols::{
-        types::{CommonPlayer, CommonResponse, GenericPlayer},
+        types::{CommonPlayer, CommonResponse, GenericPlayer, TimeoutSettings},
         valve::{self, get_optional_extracted_data, Server, ServerPlayer},
         GenericResponse,
     },
     GDErrorKind::PacketBad,
     GDResult,
-    TimeoutSettings,
 };
 use std::net::{IpAddr, SocketAddr};
