+++ crates/lib/src/lib.rs
@@ -40,7 +40,6 @@ pub mod games;
 pub mod protocols;
 #[cfg(feature = "services")]
 pub mod services;
-pub mod types;
 
 mod buffer;
 mod socket;
@@ -53,4 +52,3 @@ pub use games::*;
 pub use query::*;
 #[cfg(feature = "services")]
 pub use services::*;
-pub use types::*;
