+++ crates/lib/src/protocols/gamespy/protocols/three/protocol.rs
@@ -3,11 +3,11 @@ use byteorder::{BigEndian, LittleEndian};
 use crate::buffer::{Buffer, Utf8Decoder};
 use crate::protocols::gamespy::common::has_password;
 use crate::protocols::gamespy::three::{Player, Response, Team};
+use crate::protocols::types::TimeoutSettings;
 use crate::socket::{Socket, UdpSocket};
 use crate::utils::retry_on_timeout;
 use crate::GDErrorKind::{PacketBad, TypeParse};
-use crate::{GDErrorKind, GDResult, TimeoutSettings};
-
+use crate::{GDErrorKind, GDResult};
 use std::collections::HashMap;
 use std::net::SocketAddr;
