+++ crates/lib/src/protocols/quake/client.rs
@@ -2,10 +2,11 @@ use byteorder::LittleEndian;
 
 use crate::buffer::{Buffer, Utf8Decoder};
 use crate::protocols::quake::types::Response;
+use crate::protocols::types::TimeoutSettings;
 use crate::socket::{Socket, UdpSocket};
 use crate::utils::retry_on_timeout;
 use crate::GDErrorKind::{PacketBad, TypeParse};
-use crate::{GDErrorKind, GDResult, TimeoutSettings};
+use crate::{GDErrorKind, GDResult};
 use std::collections::HashMap;
 use std::net::SocketAddr;
 use std::slice::Iter;
