+++ crates/lib/src/protocols/quake/three.rs
@@ -1,8 +1,8 @@
 use crate::protocols::quake::client::{client_query, QuakeClient};
 use crate::protocols::quake::two::QuakeTwo;
 use crate::protocols::quake::Response;
-use crate::{GDResult, TimeoutSettings};
-
+use crate::protocols::types::TimeoutSettings;
+use crate::GDResult;
 use std::net::SocketAddr;
 use std::slice::Iter;
