+++ crates/lib/src/protocols/valve/protocol.rs
@@ -1,25 +1,27 @@
 use crate::{
     buffer::Buffer,
-    protocols::valve::{
-        types::{
-            Environment,
-            ExtraData,
-            GatheringSettings,
-            Request,
-            Response,
-            Server,
-            ServerInfo,
-            ServerPlayer,
-            TheShip,
+    protocols::{
+        types::TimeoutSettings,
+        valve::{
+            types::{
+                Environment,
+                ExtraData,
+                GatheringSettings,
+                Request,
+                Response,
+                Server,
+                ServerInfo,
+                ServerPlayer,
+                TheShip,
+            },
+            Engine,
+            ModData,
         },
-        Engine,
-        ModData,
     },
     socket::{Socket, UdpSocket},
     utils::{retry_on_timeout, u8_lower_upper},
     GDErrorKind::{BadGame, Decompress, UnknownEnumCast},
     GDResult,
-    TimeoutSettings,
 };
 
 use bzip2_rs::decoder::Decoder;
